move customer controllers to single file
