+++ src/controllers/User/createUser.ts
@@ -55,7 +55,7 @@ export const createUser = async (req: Request, res: Response) => {
     res.status(201).json({ success: true, data: user })
   } catch (error) {
     if (error instanceof Error) {
-      console.log(`Error to retrieve a treatment plan ${error.message}`)
+      console.error(`Error to create a user: ${error.message}`)
     }
 
     res
