+++ src/controllers/User/deleteUser.ts
@@ -19,7 +19,7 @@ export const deleteUser = async (req: Request, res: Response) => {
     res.status(204).send()
   } catch (error) {
     if (error instanceof Error) {
-      console.log(`Error to retrieve a treatment plan ${error.message}`)
+      console.error(`Error to delete a user: ${error.message}`)
     }
     res
       .status(500)
