+++ src/controllers/User/getUsers.ts
@@ -12,7 +12,7 @@ export const getUsers = async (_req: Request, res: Response) => {
     res.json({ success: true, data: userData })
   } catch (error) {
     if (error instanceof Error) {
-      console.log(`Error to retrieve a treatment plan ${error.message}`)
+      console.error(`Error to retrieve all users: ${error.message}`)
     }
     res
       .status(500)
