+++ src/controllers/customer.ts
@@ -4,49 +4,6 @@ import _ from 'lodash'
 import { Customer, CustomerModel } from '@schemas/customer'
 import { customerValidator } from '@validators/customer'
 
-export const getCustomer = async (req: Request, res: Response) => {
-  try {
-    const { id } = req.params
-    const customer = await CustomerModel.findById(id)
-
-    if (customer == null) {
-      return res.status(404).json({ success: false, data: null })
-    }
-
-    res.json({
-      success: true,
-      data: customer,
-    })
-  } catch (error) {
-    if (error instanceof Error) {
-      console.log(`Error to retrieve a customer ${error.message}`)
-    }
-
-    res
-      .status(500)
-      .json({ success: false, error: 'Comuniquese con su administrador' })
-  }
-}
-
-export const getCustomers = async (_req: Request, res: Response) => {
-  try {
-    const customers = await CustomerModel.find()
-
-    res.json({
-      success: true,
-      data: customers,
-    })
-  } catch (error) {
-    if (error instanceof Error) {
-      console.log(`Error to retrieve a customer ${error.message}`)
-    }
-
-    res
-      .status(500)
-      .json({ success: false, error: 'Comuniquese con su administrador' })
-  }
-}
-
 export const postCustomer = async (req: Request, res: Response) => {
   try {
     const customer = req.body as Customer
